made Printer class inline for text-ui-test
